validation: move other method checks to service
